Fix error when '${BRAND}' or '${FIRMWARE}' is empty
When the '${BRAND}' in run.sh is empty, the script will complain
'./run.sh: line 85: [: =: unary operator expected' because it becomes
'[ = "auto" ]'. Surrounding the variable with quotations solves this
problem.

And when the empty '${BRAND}' is then passed to extractor.py, it will
complain again because there is no value for argument '-b'. In fact,
if no '-b' passed to extractor.py, the script will automatically infer
the brand from the image path, so we can just omit this argument when
the '$BRAND' is empty.

Fix for '${FIRMWARE}' is the same.

Signed-off-by: ChinaNuke <chinanuke@nuke666.cn>