Fix bug caused by uncritical judgment condition
If there are errors in 'check_network', the error message will then be
captured by the 'read' command and assigned to 'PING_RESULT' as well as
'WEB_RESULT', so in this case the two 'if' statements are always true
because the strings are not empty, which in turn leads to a wrong result.

Checking if the two variables are equal to 'true' works well in this
special case.

Signed-off-by: ChinaNuke <chinanuke@nuke666.cn>